+++ src/compas_ifc/entities/extensions/IfcSite.py
@@ -1,5 +1,7 @@
 from typing import TYPE_CHECKING
 
+from compas_ifc.resources import IfcCompoundPlaneAngleMeasure_to_degrees
+
 if TYPE_CHECKING:
     from compas_ifc.entities.generated.IFC4 import IfcSite
 else:
@@ -18,3 +20,10 @@ class IfcSite(IfcSite):
     @property
     def geographic_elements(self):
         return self.children_by_type("IfcGeographicElement", recursive=True)
+
+    @property
+    def location(self):
+        if self.RefLatitude and self.RefLongitude:
+            return IfcCompoundPlaneAngleMeasure_to_degrees(self.RefLatitude), IfcCompoundPlaneAngleMeasure_to_degrees(self.RefLongitude)
+        else:
+            return None